added parseFloat to calculateTotalHours() in hourseAndStats.js in order to fix number adding problem with total_hours state

+++ server/client/src/Components/hoursAndStats.js
@@ -27,8 +27,8 @@ const HoursAndStats = () => {
   let navigate = useNavigate()
 
   const calculateTotalHours = () => {
-    const travHours = parseInt(formValues.travelHours);
-    const trainHours = parseInt(formValues.trainingHours);
+    const travHours = parseFloat(formValues.travelHours);
+    const trainHours = parseFloat(formValues.trainingHours);
     const sum = travHours + trainHours;
     return sum;
   }
